refactor(settings): move the feedback section to the Help tab
The give-feedback prompt and dialog now live under the FAQs on Help, where support-seeking users land; the Settings footer keeps updates and Discord.
